Display and update the colors when the user changes it.
TODO Right now this iterates through ALL the vertices of the model to
renew their color. The buffers are actually transferred to the GPU only
when needed, but we still iterate on all the spots here. Is there a way
to make this 'lazy'? This would require accessing the vertices at render
time...