Use a mat3 for scale and rotation of ellipsoids instead of a mat4.
We keep the translation and shape separated in the shaders, because
we often update one without the other. So we could save space by
using a mat3 for shape instead of a mat4, for which we don't use
the last column and the last row anyway.
This way we go from 16 floats to 9 floats per instance, saving memeory
and transfer time.